almost done with ornament layout generation code - smaller part of frame layout left. reverted ASHashKey back to unsigned short - when it was changed to unsigned int - no idea.

+++ libAfterBase/ashash.h
@@ -22,7 +22,7 @@ typedef struct ASHashItem
 }
 ASHashItem;
 
-typedef unsigned int ASHashKey;
+typedef unsigned short ASHashKey;
 typedef ASHashItem *ASHashBucket;
 
 typedef struct ASHashTable
@@ -31,7 +31,7 @@ typedef struct ASHashTable
   ASHashBucket *buckets;
   ASHashKey buckets_used;
   unsigned long items_num;
-  
+
   ASHashItem *most_recent ;
 
     ASHashKey (*hash_func) (ASHashableValue value, ASHashKey hash_size);
